+++ rofr/tests/test_integration.rs
@@ -116,6 +116,31 @@ async fn test_service_echo() {
 async fn test_cluster_no_services() {
     let server = nats_server::run_server("tests/nats/default.conf");
 
+    let cluster = Cluster::new(server.client_url()).unwrap();
+
+    let result = tokio::time::timeout(Duration::from_millis(50), cluster.run()).await;
+    assert!(
+        result.is_err(),
+        "cluster without services exited immediately"
+    );
+}
+
+#[service(name = "test_service_no_endpoints", version = "0.1.2")]
+trait TestServiceNoEndpoints {
+    type Context;
+}
+
+#[derive(Debug)]
+struct TestServiceNoEndpointsImpl;
+
+impl TestServiceNoEndpoints for TestServiceNoEndpointsImpl {
+    type Context = ();
+}
+
+#[tokio::test]
+async fn test_serivce_no_endpoints() {
+    let server = nats_server::run_server("tests/nats/default.conf");
+
     let mut cluster = Cluster::new(server.client_url()).unwrap();
     let test_service = TestServiceNoEndpointsImpl::service(());
     cluster.register(test_service);