+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Added idling indefinitely when a cluster is run without any services
   registered.
+- Added idling indefinitely when a service doesn't contain any endpoints or
+  streams.
 
 ## v0.1.1